Add the different ordering and the edit event
modified:   views/admin/tags.jade
With the <th> field in the TagList, use html5 data-* to discover the
comparator, remove the _id in table, add lastModified and lastModifier
field to the TagList.

modified:   public/javascripts/tags.js
Modified the order function in the tagView, let this function to handle
different comparator. Add a event for editting tag.